Rearrange files
This moves suite config handling code into the library proper instead of
the script, and creates a subdir for suite configs.
